Pull up method: Traps Child class_Less code duplication
적용 리팩토링 또는 패턴
- Pull up method

<br>

대상
- AlarmTrap.java, DisintegrationTrap.java, DistortionTrap.java, GrimTrap.java, GuardianTrap.java, SummoningTrap.java, WormDartTrap.java

<br>

수행 이유
- Code duplicate

<br>

효과
- Less code duplication
