--- core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/traps/MobsRespawnPoints.java
@@ -1,38 +1,0 @@
-package com.shatteredpixel.shatteredpixeldungeon.levels.traps;
-
-import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
-import com.shatteredpixel.shatteredpixeldungeon.actors.Actor;
-import com.watabou.utils.PathFinder;
-import com.watabou.utils.Random;
-
-import java.util.ArrayList;
-
-public interface MobsRespawnPoints {
-    static ArrayList<Integer> getMobsRespawnPoints(int nMobs, int pos) {
-        if (Random.Int( 2 ) == 0) {
-            nMobs++;
-            if (Random.Int( 2 ) == 0) {
-                nMobs++;
-            }
-        }
-
-        ArrayList<Integer> candidates = new ArrayList<>();
-
-        for (int i = 0; i < PathFinder.NEIGHBOURS8.length; i++) {
-            int p = pos + PathFinder.NEIGHBOURS8[i];
-            if (Actor.findChar( p ) == null && (Dungeon.level.passable[p] || Dungeon.level.avoid[p])) {
-                candidates.add( p );
-            }
-        }
-
-        ArrayList<Integer> respawnPoints = new ArrayList<>();
-
-        while (nMobs > 0 && candidates.size() > 0) {
-            int index = Random.index( candidates );
-
-            respawnPoints.add( candidates.remove( index ) );
-            nMobs--;
-        }
-        return respawnPoints;
-    }
-}
